added proper error handling
